// course/src/types.ts
export interface HeaderProp {
  courseName: string;
}

export interface ContentProps {
  courseParts: CoursePart[];
}

interface CoursePartBase {
  name: string;
  exerciseCount: number;
}

interface CoursePartDescription extends CoursePartBase{
  description: string
}

interface CoursePartBasic extends CoursePartDescription{
  kind: "basic";
}

interface CoursePartGroup extends CoursePartBase{
  groupProjectCount: number;
  kind: "group";
}

interface CoursePartBackground extends CoursePartDescription{
  backgroundMaterial: string;
  kind: "background";
}

interface CoursePartSpecial extends CoursePartDescription{
  requirements: string[]
  kind: "special"
}

export type CoursePart = CoursePartBasic | CoursePartGroup | CoursePartBackground | CoursePartSpecial

export interface TotalProp {
  totalExercise: number
}